requirements.txt
# Scraping
requests==2.32.3
beautifulsoup4==4.12.3

# PDF parsing
pymupdf==1.25.0

# Ingestion & RAG (Updating to 0.3.1 for better Python compatibility)
langchain==0.3.2
langchain-community==0.3.2
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langchain-core==0.3.9

# Vector store
chromadb==0.5.15

# Interface
streamlit==1.39.0

# Utilitaires
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.66.5
pandas==2.2.3
pydantic>=2.9.0
